add api framework, fauna migrations, and auth
closes #5

This is a bit of a monster commit which adds basic auth functionality.
To do so it also implements an API framework, and a fauna database
model.

We only interact with Fauna through UDFs each of which can be typed
manually. All resources, like collections and indexes in addition to
functions,  export their names as constants, which are used in other
resources to ensure they are properly identified.

Each endpoint is defined as a function which returns an object which can
extract the type of. The main worker script handles running the correct
endpoint, validating inputs, and fitting the returned object into an
HTTP response .