tweak a bunch of collision detection stuff for collections
When dragging into a collection from another room, rects were not being
updated properly. This was due to a bug in combinedRefs I think.

I also added a config to the droppables to triggers recalculation of all
rects when something resizes. It might be a lil jank, but should work!