mutate indirectly to handle ios safari image quirks
When uploading an image in ios safari it prompts you to take a picture.
if you do it seems to put that event loop into a secure mode, which
means we can't access indexedDB, which means the call to mutate
replicache fails. To avoid this we setState in the event loop, and then
use an effect in  react to actually write it replicache.