migrate space data to supabase
This is a rather big change motivated by the upcoming studio work. Right
now, in order to have the data available in the local replicache
instance, we have a bunch of functions for syncing space metadata (like
title, spaceid, description, etc) to all the spaces that might need it.

This is quite a complicated system and bugs have turned up before
because we didn't take it into account properly.

Instead now, we just store the data in supabase and fetch it on page
load and via useSWR hooks. This makes local mutation more complicated,
but only slightly, and gives us greater freedom when adding new flows
around spaces.