update replicache and refactor presence architecture
Use supabase sockets instead of DO so that we can use their client
library and re-use it for other realtime needs, i.e notifications.

This means we can't clear facts on disconnect but that was unreliable
anyway, so instead we now do a keep alive kind of thing, and clear out
old ephemeralfacts.